--- comically/src/image.rs
@@ -1,954 +1,0 @@
-use anyhow::{Context, Result};
-use arrayvec::ArrayVec;
-use fast_image_resize as fr;
-use fr::images::Image as FrImage;
-use imageproc::image::{
-    imageops, load_from_memory, ColorType, DynamicImage, GenericImageView, GrayImage, Luma,
-    SubImage,
-};
-use imageproc::stats::histogram;
-use rayon::iter::ParallelIterator;
-use rayon::slice::ParallelSlice;
-use webp::WebPMemory;
-
-use crate::archive::ArchiveFile;
-use crate::comic::{ComicConfig, ProcessedImage, SplitStrategy};
-
-#[derive(Debug, Clone, Copy, PartialEq, serde::Serialize, serde::Deserialize)]
-pub enum PngCompression {
-    Fast,
-    Default,
-    Best,
-}
-
-impl PngCompression {
-    pub fn cycle(&self) -> Self {
-        match self {
-            PngCompression::Fast => PngCompression::Default,
-            PngCompression::Default => PngCompression::Best,
-            PngCompression::Best => PngCompression::Fast,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, serde::Serialize, serde::Deserialize)]
-pub enum ImageFormat {
-    Jpeg { quality: u8 },
-    Png { compression: PngCompression },
-    WebP { quality: u8 },
-}
-
-impl ImageFormat {
-    pub fn cycle(&self) -> Self {
-        match self {
-            ImageFormat::Jpeg { .. } => ImageFormat::Png {
-                compression: PngCompression::Default,
-            },
-            ImageFormat::Png { .. } => ImageFormat::WebP { quality: 85 },
-            ImageFormat::WebP { .. } => ImageFormat::Jpeg { quality: 85 },
-        }
-    }
-
-    pub fn extension(&self) -> &'static str {
-        match self {
-            ImageFormat::Jpeg { .. } => "jpg",
-            ImageFormat::Png { .. } => "png",
-            ImageFormat::WebP { .. } => "webp",
-        }
-    }
-
-    pub fn adjust_quality(&mut self, increase: bool, fine: bool) {
-        let step = if fine { 1 } else { 5 };
-        match self {
-            ImageFormat::Jpeg { quality } | ImageFormat::WebP { quality } => {
-                if increase {
-                    *quality = (*quality + step).min(100);
-                } else {
-                    *quality = quality.saturating_sub(step);
-                }
-            }
-            ImageFormat::Png { compression } => {
-                *compression = if increase {
-                    match compression {
-                        PngCompression::Fast => PngCompression::Default,
-                        PngCompression::Default => PngCompression::Best,
-                        PngCompression::Best => PngCompression::Best,
-                    }
-                } else {
-                    match compression {
-                        PngCompression::Fast => PngCompression::Fast,
-                        PngCompression::Default => PngCompression::Fast,
-                        PngCompression::Best => PngCompression::Default,
-                    }
-                };
-            }
-        }
-    }
-}
-
-pub struct Split<T>(ArrayVec<T, 3>);
-
-impl<T> Split<T> {
-    #[inline(always)]
-    pub fn one(t: T) -> Self {
-        let mut vec = ArrayVec::new();
-        vec.push(t);
-        Split(vec)
-    }
-
-    #[inline(always)]
-    pub fn two(t1: T, t2: T) -> Self {
-        let mut vec = ArrayVec::new();
-        vec.push(t1);
-        vec.push(t2);
-        Split(vec)
-    }
-
-    #[inline(always)]
-    pub fn three(t1: T, t2: T, t3: T) -> Self {
-        Split(ArrayVec::from([t1, t2, t3]))
-    }
-
-    #[inline(always)]
-    pub fn map<U, F: FnMut(T) -> U>(self, f: F) -> Split<U> {
-        Split(self.0.into_iter().map(f).collect())
-    }
-}
-
-impl<T> IntoIterator for Split<T> {
-    type Item = T;
-    type IntoIter = <ArrayVec<T, 3> as IntoIterator>::IntoIter;
-
-    #[inline(always)]
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.into_iter()
-    }
-}
-
-pub fn process_archive_images(
-    archive: impl Iterator<Item = anyhow::Result<ArchiveFile>> + Send,
-    config: &ComicConfig,
-) -> Result<Vec<ProcessedImage>> {
-    log::info!("Processing archive images");
-
-    // 1. Collect archive files (serial, fast - no parallelism overhead)
-    let files: Vec<ArchiveFile> = archive
-        .filter_map(|result| {
-            result
-                .map_err(|e| log::warn!("Failed to load archive file: {}", e))
-                .ok()
-        })
-        .collect();
-
-    log::debug!("Loaded {} files from archive", files.len());
-
-    // Calculate chunk size to minimize rayon overhead
-    // Use larger chunks to reduce coordination overhead
-    let num_threads = rayon::current_num_threads();
-    let chunk_size = (files.len() / num_threads).max(1);
-
-    log::debug!(
-        "Processing {} files with {} threads, chunk size: {}",
-        files.len(),
-        num_threads,
-        chunk_size
-    );
-
-    // 2. Single parallel stage: decode + process + encode
-    // This eliminates intermediate Vec allocation and keeps data hot in cache
-    let mut images: Vec<ProcessedImage> = files
-        .par_chunks(chunk_size)
-        .flat_map_iter(|chunk| {
-            chunk.iter().flat_map(|archive_file| {
-                let mut encoded_images = ArrayVec::<ProcessedImage, 3>::new();
-
-                // Decode image
-                let img = match load_from_memory(&archive_file.data) {
-                    Ok(img) => img,
-                    Err(e) => {
-                        log::warn!(
-                            "Failed to decode {}: {}",
-                            archive_file.file_name.display(),
-                            e
-                        );
-                        return encoded_images;
-                    }
-                };
-
-                // Process image (transform, crop, resize, split)
-                let processed_images = process(img, config);
-
-                // Encode immediately while data is hot in cache
-                for (i, img) in processed_images.into_iter().enumerate() {
-                    match encode_image_part(archive_file, &img, i, config.image_format) {
-                        Ok(processed) => encoded_images.push(processed),
-                        Err(e) => {
-                            log::warn!(
-                                "Failed to encode {}: {}",
-                                archive_file.file_name.display(),
-                                e
-                            );
-                        }
-                    }
-                }
-                encoded_images
-            })
-        })
-        .collect();
-
-    // 4. Serial sort + dedup (fast, no benefit from parallelism)
-    images.sort_unstable_by(|a, b| a.file_name.cmp(&b.file_name));
-    images.dedup_by(|a, b| a.file_name == b.file_name);
-
-    Ok(images)
-}
-
-/// Process a single image file with Kindle-optimized transformations
-pub fn process(img: DynamicImage, config: &ComicConfig) -> Split<DynamicImage> {
-    let img = transform(img.into_luma8(), config.brightness, config.gamma);
-
-    let gray_images = if config.auto_crop {
-        if let Some(cropped) = auto_crop(&img) {
-            process_image_view(&cropped.to_image(), config)
-        } else {
-            process_image_view(&img, config)
-        }
-    } else {
-        process_image_view(&img, config)
-    };
-
-    // Convert GrayImage to DynamicImage
-    gray_images.map(DynamicImage::ImageLuma8)
-}
-
-fn process_image_view(img: &GrayImage, c: &ComicConfig) -> Split<GrayImage> {
-    let target = c.device_dimensions();
-    let (width, height) = img.dimensions();
-    let is_double_page = width > height;
-
-    let margin = c.margin_color;
-
-    match c.split {
-        SplitStrategy::None => {
-            // Just resize, no splitting or rotation
-            Split::one(resize_image(img, target, margin))
-        }
-        SplitStrategy::Split => {
-            if is_double_page {
-                // Split double pages
-                let (left, right) = split_double_pages(img);
-
-                let (left_resized, right_resized) = rayon::join(
-                    || resize_image(&left, target, margin),
-                    || resize_image(&right, target, margin),
-                );
-
-                // Determine order based on right_to_left setting
-                let (first, second) = if c.right_to_left {
-                    (right_resized, left_resized)
-                } else {
-                    (left_resized, right_resized)
-                };
-
-                Split::two(first, second)
-            } else {
-                Split::one(resize_image(img, target, margin))
-            }
-        }
-        SplitStrategy::Rotate => {
-            if is_double_page {
-                let rotated = rotate_image_90(img, c.right_to_left);
-                Split::one(resize_image(&rotated, target, margin))
-            } else {
-                Split::one(resize_image(img, target, margin))
-            }
-        }
-        SplitStrategy::RotateAndSplit => {
-            if is_double_page {
-                let (left, right) = split_double_pages(img);
-
-                let mut rotated_resized = None;
-                let mut left_resized = None;
-                let mut right_resized = None;
-
-                rayon::scope(|s| {
-                    s.spawn(|_| {
-                        let rotated = rotate_image_90(img, c.right_to_left);
-                        rotated_resized = Some(resize_image(&rotated, target, margin));
-                    });
-                    s.spawn(|_| {
-                        left_resized = Some(resize_image(&left, target, margin));
-                    });
-                    s.spawn(|_| {
-                        right_resized = Some(resize_image(&right, target, margin));
-                    });
-                });
-
-                let rotated_resized = rotated_resized.unwrap();
-                let left_resized = left_resized.unwrap();
-                let right_resized = right_resized.unwrap();
-
-                let (first, second) = if c.right_to_left {
-                    (right_resized, left_resized)
-                } else {
-                    (left_resized, right_resized)
-                };
-
-                Split::three(rotated_resized, first, second)
-            } else {
-                Split::one(resize_image(img, target, margin))
-            }
-        }
-    }
-}
-
-/// Gamma correction lookup table
-/// Only computed once per unique gamma value (256 iterations) to avoid slow float operations
-static GAMMA_LUT: std::sync::OnceLock<[u8; 256]> = std::sync::OnceLock::new();
-
-fn gamma_lut(gamma: f32) -> &'static [u8; 256] {
-    GAMMA_LUT.get_or_init(|| {
-        let mut lut = [0u8; 256];
-        for (i, pixel) in lut.iter_mut().enumerate() {
-            let normalized = i as f32 / 255.0;
-            let corrected = normalized.powf(gamma);
-            *pixel = (corrected * 255.0).round().clamp(0.0, 255.0) as u8;
-        }
-        lut
-    })
-}
-
-/// gamma - 0.1 to 3.0, where 1.0 = no change, <1 = brighter, >1 = more contrast
-fn transform(mut img: GrayImage, brightness: i32, gamma: f32) -> GrayImage {
-    let gamma = gamma.clamp(0.1, 3.0);
-    // only apply gamma if it's not 1.0
-    if (gamma - 1.0).abs() > 0.01 {
-        imageproc::map::map_colors_mut(&mut img, |pixel| {
-            Luma([gamma_lut(gamma)[pixel[0] as usize]])
-        });
-    }
-
-    // Apply autocontrast - find actual min/max and stretch to 0-255
-    let hist = histogram(&img);
-
-    let channel_hist = &hist.channels[0];
-
-    let min = channel_hist
-        .iter()
-        .position(|&count| count > 0)
-        .unwrap_or(0) as u8;
-    let max = channel_hist
-        .iter()
-        .rposition(|&count| count > 0)
-        .unwrap_or(255) as u8;
-
-    // Only stretch if there's a range to work with
-    if max > min {
-        img = imageproc::contrast::stretch_contrast(&img, min, max, 0, 255);
-    }
-
-    // Only apply manual adjustments if explicitly set
-    if brightness != 0 {
-        imageops::colorops::brighten_in_place(&mut img, brightness);
-    }
-
-    img
-}
-
-fn split_double_pages(img: &GrayImage) -> (GrayImage, GrayImage) {
-    let (width, height) = img.dimensions();
-
-    let left = imageops::crop_imm(img, 0, 0, width / 2, height).to_image();
-    let right = imageops::crop_imm(img, width / 2, 0, width / 2, height).to_image();
-
-    (left, right)
-}
-
-fn rotate_image_90(img: &GrayImage, clockwise: bool) -> GrayImage {
-    let (width, height) = img.dimensions();
-    let mut rotated = GrayImage::new(height, width);
-
-    for y in 0..height {
-        for x in 0..width {
-            let pixel = img.get_pixel(x, y);
-            if clockwise {
-                rotated.put_pixel(height - 1 - y, x, *pixel);
-            } else {
-                rotated.put_pixel(y, width - 1 - x, *pixel);
-            }
-        }
-    }
-
-    rotated
-}
-
-// Pixel values above this are considered "white"
-const WHITE_THRESHOLD: u8 = 230;
-// Minimum width to consider cropping
-const MIN_MARGIN_WIDTH: u32 = 10;
-// Extra margin to keep, avoiding cutting content
-const SAFETY_MARGIN: u32 = 2;
-
-/// Auto-crop white margins from all sides of the image
-fn auto_crop(img: &GrayImage) -> Option<SubImage<&GrayImage>> {
-    let (width, height) = img.dimensions();
-
-    // Left margin: scan from left to right
-    let mut left_margin = 0;
-    'left: for x in 0..width {
-        for y in 0..height {
-            if img.get_pixel(x, y)[0] < WHITE_THRESHOLD && is_not_noise(img, x, y) {
-                left_margin = x;
-                break 'left;
-            }
-        }
-    }
-
-    // Right margin: scan from right to left
-    let mut right_margin = width - 1;
-    'right: for x in (0..width).rev() {
-        for y in 0..height {
-            if img.get_pixel(x, y)[0] < WHITE_THRESHOLD && is_not_noise(img, x, y) {
-                right_margin = x;
-                break 'right;
-            }
-        }
-    }
-
-    // Top margin: scan from top to bottom
-    let mut top_margin = 0;
-    'top: for y in 0..height {
-        for x in 0..width {
-            if img.get_pixel(x, y)[0] < WHITE_THRESHOLD && is_not_noise(img, x, y) {
-                top_margin = y;
-                break 'top;
-            }
-        }
-    }
-
-    // Bottom margin: scan from bottom to top
-    let mut bottom_margin = height - 1;
-    'bottom: for y in (0..height).rev() {
-        for x in 0..width {
-            if img.get_pixel(x, y)[0] < WHITE_THRESHOLD && is_not_noise(img, x, y) {
-                bottom_margin = y;
-                break 'bottom;
-            }
-        }
-    }
-
-    // If we didn't find any content, return None
-    if left_margin >= right_margin || top_margin >= bottom_margin {
-        return None;
-    }
-
-    // Apply safety margin
-    left_margin = left_margin.saturating_sub(SAFETY_MARGIN);
-    right_margin = (right_margin + SAFETY_MARGIN).min(width - 1);
-    top_margin = top_margin.saturating_sub(SAFETY_MARGIN);
-    bottom_margin = (bottom_margin + SAFETY_MARGIN).min(height - 1);
-
-    let crop_width = right_margin.saturating_sub(left_margin).saturating_add(1);
-    let crop_height = bottom_margin.saturating_sub(top_margin).saturating_add(1);
-
-    let left_margin_size = left_margin;
-    let right_margin_size = width.saturating_sub(right_margin).saturating_sub(1);
-    let top_margin_size = top_margin;
-    let bottom_margin_size = height.saturating_sub(bottom_margin).saturating_sub(1);
-
-    // Only crop if at least one margin is wide enough AND we have positive crop dimensions
-    let should_crop_horizontal = (left_margin_size >= MIN_MARGIN_WIDTH
-        || right_margin_size >= MIN_MARGIN_WIDTH)
-        && crop_width > 0
-        && crop_width < width;
-    let should_crop_vertical = (top_margin_size >= MIN_MARGIN_WIDTH
-        || bottom_margin_size >= MIN_MARGIN_WIDTH)
-        && crop_height > 0
-        && crop_height < height;
-
-    if should_crop_horizontal || should_crop_vertical {
-        return Some(imageops::crop_imm(
-            img,
-            left_margin,
-            top_margin,
-            crop_width,
-            crop_height,
-        ));
-    }
-
-    None
-}
-
-/// Check if a pixel is likely to be content rather than noise
-#[inline]
-fn is_not_noise(img: &GrayImage, x: u32, y: u32) -> bool {
-    // how many neighbors need to be dark to consider the pixel content
-    const REQUIRED_NEIGHBORS: i32 = 3;
-    // Look a bit farther for connected pixels
-    const DISTANCE: i32 = 4;
-
-    let (width, height) = img.dimensions();
-    let mut dark_neighbors = 0;
-
-    for dy in -DISTANCE..=DISTANCE {
-        for dx in -DISTANCE..=DISTANCE {
-            // skip center
-            if dx == 0 && dy == 0 {
-                continue;
-            }
-
-            let nx = x as i32 + dx;
-            let ny = y as i32 + dy;
-
-            // Make sure coordinates are valid
-            if nx >= 0
-                && ny >= 0
-                && nx < width as i32
-                && ny < height as i32
-                && img.get_pixel(nx as u32, ny as u32)[0] < WHITE_THRESHOLD
-            {
-                dark_neighbors += 1;
-
-                // Early return if we have enough neighbors
-                if dark_neighbors >= REQUIRED_NEIGHBORS {
-                    return true;
-                }
-            }
-        }
-    }
-
-    dark_neighbors >= REQUIRED_NEIGHBORS
-}
-
-fn resize_image(
-    img: &GrayImage,
-    device_dimensions: (u32, u32),
-    margin_color: Option<u8>,
-) -> GrayImage {
-    let (target_width, target_height) = device_dimensions;
-    let (width, height) = img.dimensions();
-
-    // Calculate aspect-fit dimensions
-    let width_ratio = target_width as f32 / width as f32;
-    let height_ratio = target_height as f32 / height as f32;
-    let ratio = width_ratio.min(height_ratio);
-
-    let new_width = (width as f32 * ratio) as u32;
-    let new_height = (height as f32 * ratio) as u32;
-
-    // Choose algorithm based on scaling direction
-    let algorithm = if ratio < 1.0 {
-        // Downscaling: Lanczos3 preserves detail
-        fr::ResizeAlg::Convolution(fr::FilterType::Lanczos3)
-    } else {
-        // Upscaling: CatmullRom gives smoother results
-        fr::ResizeAlg::Convolution(fr::FilterType::CatmullRom)
-    };
-
-    // Create source image view (need to clone data since fast_image_resize requires mutable slice)
-    let src_buffer = img.as_raw().clone();
-    let src_image = FrImage::from_vec_u8(width, height, src_buffer, fr::PixelType::U8).unwrap();
-
-    // Create destination buffer
-    let mut dst_buffer = vec![0u8; (new_width * new_height) as usize];
-    let mut dst_image =
-        FrImage::from_slice_u8(new_width, new_height, &mut dst_buffer, fr::PixelType::U8).unwrap();
-
-    // Perform resize
-    let mut resizer = fr::Resizer::new();
-    resizer
-        .resize(
-            &src_image,
-            &mut dst_image,
-            Some(&fr::ResizeOptions::new().resize_alg(algorithm)),
-        )
-        .unwrap();
-
-    // Convert back to GrayImage
-    let resized = GrayImage::from_raw(new_width, new_height, dst_buffer).unwrap();
-
-    // If exact fit, return as-is
-    if new_width == target_width && new_height == target_height {
-        return resized;
-    }
-
-    // Add margins if requested
-    match margin_color {
-        Some(color) => {
-            let mut result = GrayImage::from_pixel(target_width, target_height, Luma([color]));
-            let x_offset = (target_width - new_width) / 2;
-            let y_offset = (target_height - new_height) / 2;
-            imageops::overlay(&mut result, &resized, x_offset.into(), y_offset.into());
-            result
-        }
-        None => resized,
-    }
-}
-
-/// Compress an image to JPEG format with the specified quality
-pub fn compress_to_jpeg<W>(img: &DynamicImage, writer: &mut W, quality: u8) -> Result<()>
-where
-    W: std::io::Write,
-{
-    let mut encoder =
-        imageproc::image::codecs::jpeg::JpegEncoder::new_with_quality(writer, quality);
-
-    encoder
-        .encode_image(img)
-        .with_context(|| "Failed to compress image to JPEG")?;
-
-    Ok(())
-}
-
-/// Compress an image to PNG format with the specified compression level
-pub fn compress_to_png<W>(
-    img: &DynamicImage,
-    writer: &mut W,
-    compression: PngCompression,
-) -> Result<()>
-where
-    W: std::io::Write,
-{
-    use imageproc::image::codecs::png::{CompressionType, FilterType, PngEncoder};
-    use imageproc::image::ImageEncoder;
-
-    let compression_type = match compression {
-        PngCompression::Fast => CompressionType::Fast,
-        PngCompression::Default => CompressionType::Default,
-        PngCompression::Best => CompressionType::Best,
-    };
-
-    let is_grayscale = img.color() == ColorType::L8 || img.color() == ColorType::La8;
-
-    let encoder = PngEncoder::new_with_quality(
-        writer,
-        compression_type,
-        if is_grayscale {
-            FilterType::NoFilter
-        } else {
-            FilterType::Adaptive
-        },
-    );
-
-    encoder
-        .write_image(
-            img.as_bytes(),
-            img.width(),
-            img.height(),
-            img.color().into(),
-        )
-        .with_context(|| "Failed to compress image to PNG")?;
-
-    Ok(())
-}
-
-/// Compress an image to WebP format with the specified quality
-pub fn compress_to_webp(img: &DynamicImage, quality: u8) -> Result<WebPMemory> {
-    let img = DynamicImage::from(img.to_rgb8());
-    let encoder = webp::Encoder::from_image(&img)
-        .map_err(|e| anyhow::anyhow!("Failed to create WebP encoder: {}", e))?;
-    let webp_data = encoder.encode(quality as f32);
-    Ok(webp_data)
-}
-
-fn encode_image_part(
-    original: &ArchiveFile,
-    img: &DynamicImage,
-    part_num: usize,
-    format: ImageFormat,
-) -> Result<ProcessedImage> {
-    let file_name = {
-        let file = original.parent().display();
-        let stem = original.file_stem().to_string_lossy();
-        let extension = format.extension();
-        format!("{file}_{stem}_{part_num:03}.{extension}")
-    };
-
-    let dimensions = img.dimensions();
-
-    encode_image(img, &format)
-        .map(|data| ProcessedImage {
-            file_name,
-            data,
-            dimensions,
-            format,
-        })
-        .inspect(|processed| {
-            log::trace!("Encoded image: {}", processed.file_name);
-        })
-}
-
-fn encode_image(img: &DynamicImage, format: &ImageFormat) -> Result<Vec<u8>> {
-    let mut buffer = Vec::new();
-
-    match format {
-        ImageFormat::Jpeg { quality } => {
-            compress_to_jpeg(img, &mut buffer, *quality)?;
-        }
-        ImageFormat::Png { compression } => {
-            compress_to_png(img, &mut buffer, *compression)?;
-        }
-        ImageFormat::WebP { quality } => {
-            let webp_data = compress_to_webp(img, *quality)?;
-            buffer.extend_from_slice(&webp_data);
-        }
-    }
-
-    Ok(buffer)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use imageproc::image::{GrayImage, Luma};
-
-    #[test]
-    fn test_basic_cropping() {
-        // Create a 100x50 image with 25px left margin and 35px right margin
-        let test_img = create_test_image(100, 50, 25, 35, &[]);
-
-        let result = auto_crop(&test_img);
-        assert!(result.is_some(), "Cropping should have succeeded");
-
-        let cropped = result.unwrap();
-        let (cropped_width, cropped_height) = cropped.dimensions();
-
-        assert_eq!(cropped_height, 50);
-        assert_eq!(
-            cropped_width,
-            100 - 60 + 2 * SAFETY_MARGIN,
-            "cropped width should be 60"
-        );
-    }
-
-    #[test]
-    fn test_noise_resilience() {
-        // Create an image with proper margins and noise in those margins
-        let test_img = create_test_image(
-            100,
-            50,
-            20,
-            20,
-            &[
-                // left
-                (5, 1),
-                (5, 2),
-                (5, 3),
-                // right
-                (95, 1),
-                (95, 2),
-                (95, 3),
-            ],
-        );
-
-        let result = auto_crop(&test_img);
-
-        let cropped = result.expect("Cropping should have succeeded");
-        let (cropped_width, _) = cropped.dimensions();
-
-        assert_eq!(cropped_width, 100 - 40 + 2 * SAFETY_MARGIN,);
-    }
-
-    #[test]
-    fn test_no_margins() {
-        let test_img = create_test_image(100, 50, 0, 0, &[]);
-
-        let cropped_option = auto_crop(&test_img);
-
-        assert!(
-            cropped_option.is_none(),
-            "Expected None for image without margins"
-        );
-    }
-
-    #[test]
-    fn test_large_margin() {
-        let test_img = create_test_image(100, 50, 30, 5, &[]);
-
-        let result = auto_crop(&test_img);
-        assert!(
-            result.is_some(),
-            "Expected cropping to succeed with one large margin"
-        );
-
-        let (width, height) = result.unwrap().dimensions();
-        assert_eq!(width, 100 - 35 + 2 * SAFETY_MARGIN);
-        assert_eq!(height, 50);
-    }
-
-    #[test]
-    fn test_complex_image() {
-        // The shape has a 40px margin on both sides
-        let mut img = GrayImage::new(200, 100);
-
-        for y in 0..100 {
-            for x in 0..200 {
-                img.put_pixel(x, y, Luma([255]));
-            }
-        }
-
-        for y in 20..80 {
-            for x in 40..160 {
-                img.put_pixel(x, y, Luma([0]));
-            }
-        }
-
-        // Apply cropping
-        let result = auto_crop(&img);
-
-        let cropped = result.expect("Cropping should have succeeded");
-        let (cropped_width, _) = cropped.dimensions();
-
-        assert_eq!(
-            cropped_width,
-            200 - 80 + 2 * SAFETY_MARGIN,
-            "invalid cropped width"
-        );
-    }
-
-    #[test]
-    fn test_vertical_cropping() {
-        // Create a 100x100 image with 20px top margin and 30px bottom margin
-        let test_img = create_test_image_with_vertical(100, 100, 0, 0, 20, 30, &[]);
-
-        let top = find_topmost_content(&test_img);
-        let bottom = find_bottommost_content(&test_img);
-
-        assert_eq!(top, 20, "Top content should start at y=20");
-        assert_eq!(bottom, 69, "Bottom content should end at y=69");
-
-        let result = auto_crop(&test_img);
-        assert!(result.is_some(), "Cropping should have succeeded");
-
-        let cropped = result.unwrap();
-        let (_, cropped_height) = cropped.dimensions();
-
-        assert_eq!(
-            cropped_height,
-            100 - 50 + 2 * SAFETY_MARGIN,
-            "Cropped height {} is outside expected range [45-55]",
-            cropped_height
-        );
-    }
-
-    #[test]
-    fn test_all_margins_cropping() {
-        // Create a 200x200 image with margins on all sides
-        let test_img = create_test_image_with_vertical(200, 200, 30, 40, 25, 35, &[]);
-
-        let result = auto_crop(&test_img);
-        assert!(result.is_some(), "Cropping should have succeeded");
-
-        let cropped = result.unwrap();
-        let (cropped_width, cropped_height) = cropped.dimensions();
-
-        assert_eq!(
-            cropped_width,
-            200 - 70 + 2 * SAFETY_MARGIN,
-            "Cropped width {} is outside expected range [125-135]",
-            cropped_width
-        );
-        assert_eq!(
-            cropped_height,
-            200 - 60 + 2 * SAFETY_MARGIN,
-            "Cropped height {} is outside expected range [135-145]",
-            cropped_height
-        );
-    }
-
-    /// Create a test image with known margins and content
-    fn create_test_image(
-        width: u32,
-        height: u32,
-        left_margin: u32,
-        right_margin: u32,
-        noise_positions: &[(u32, u32)],
-    ) -> GrayImage {
-        let mut img = GrayImage::new(width, height);
-
-        // Fill with white
-        for y in 0..height {
-            for x in 0..width {
-                img.put_pixel(x, y, Luma([255]));
-            }
-        }
-
-        // Add content (black area) in the middle
-        for y in 0..height {
-            let right_boundary = width.saturating_sub(right_margin);
-            for x in left_margin..right_boundary {
-                img.put_pixel(x, y, Luma([0]));
-            }
-        }
-
-        // Add noise pixels
-        for &(x, y) in noise_positions {
-            img.put_pixel(x, y, Luma([0]));
-        }
-
-        img
-    }
-
-    // Helper function for creating test images with top and bottom margins
-    fn create_test_image_with_vertical(
-        width: u32,
-        height: u32,
-        left_margin: u32,
-        right_margin: u32,
-        top_margin: u32,
-        bottom_margin: u32,
-        noise_positions: &[(u32, u32)],
-    ) -> GrayImage {
-        let mut img = GrayImage::new(width, height);
-
-        // Fill with white
-        for y in 0..height {
-            for x in 0..width {
-                img.put_pixel(x, y, Luma([255]));
-            }
-        }
-
-        // Add content (black area) in the middle
-        let right_boundary = width.saturating_sub(right_margin);
-        let bottom_boundary = height.saturating_sub(bottom_margin);
-
-        for y in top_margin..bottom_boundary {
-            for x in left_margin..right_boundary {
-                img.put_pixel(x, y, Luma([0]));
-            }
-        }
-
-        // Add noise pixels
-        for &(x, y) in noise_positions {
-            if x < width && y < height {
-                img.put_pixel(x, y, Luma([0]));
-            }
-        }
-
-        img
-    }
-
-    fn find_topmost_content(img: &GrayImage) -> u32 {
-        let (width, height) = img.dimensions();
-        for y in 0..height {
-            for x in 0..width {
-                if img.get_pixel(x, y)[0] < 230 {
-                    return y;
-                }
-            }
-        }
-        0
-    }
-
-    fn find_bottommost_content(img: &GrayImage) -> u32 {
-        let (width, height) = img.dimensions();
-        for y in (0..height).rev() {
-            for x in 0..width {
-                if img.get_pixel(x, y)[0] < 230 {
-                    return y;
-                }
-            }
-        }
-        height - 1
-    }
-}
